feat(params-ex): Add WriteOnce and summary for interface
This commit primarily adds detailed comments on the IParamsClientEx interface which explain the role of each function and property within the interface.

The ‘Write’ method signatures are also updated in both IParamsClient and ParamsClient, and a corresponding implementation, WriteOnce, is added in ParamsClientEx. The purpose of these changes is to enable the function to return a resultant value after parameter modifications.

Asana: https://app.asana.com/0/1203851531040615/1204913862486908/f