fix: add AvoidAdsb flight mode for ArdupilotCopterMode
A new flight mode called 'AvoidAdsb' is added to ArdupilotCopterMode, which is designed around ADS-B protocol for avoiding manned aircraft. This mode should not be set as a pilot selectable flight mode. The descriptions for this mode are properly localized and included in resource files 'RS.resx' and 'RS.ru.resx'. Reordered the configuration of modes in ArdupilotPlaneMode and ArdupilotCopterMode to include well-known modes at the beginning of the array.

Asana: https://app.asana.com/0/1203851531040615/1204961023189355/f