feat(vehicle-mode): add 'InternalMode' flag to VehicleMode
This commit introduces an 'InternalMode' flag to 'VehicleMode', 'ArdupilotCopterMode', and 'ArdupilotPlaneMode' classes. The flag helps determine if a mode is internal or user-settable in the client interfaces. 'InternalMode' is expected not to be visible in the user interface. Also, a comment is corrected for better comprehension.

Asana: https://app.asana.com/0/1203851531040615/1204961023189355/f