feat(sdr-store): refactor file storage system and add file cache
This commit was a refactoring of the file storage system from Mavlink and the inclusion of a new file cache system. Firstly, file types related to ASV SDR have been moved to their own classes to make it clearer where those classes should be used.

The introduction of a file cache system comes with many advantages. By reusing the same file set, we can reduce the overhead of creating new file objects and discarding old ones. This change could potentially increase the performance of the system by lessening the number of file read/writes that need to take place.

Additionally, File Rename and Move operations became safer as files that are currently in use will not be processed. This ensures the integrity of file operation processes.

Lastly, a variety of useful logging information was added for debugging and traceability purposes. It offers easier tracking of events such as renaming, moving, or deleting files.