feat(mavlink): add ROI mission item to IMissionClientEx interface
In this update, a function was added to the IMissionClientEx interface to construct and add a Region of Interest (RoI) mission item. This function takes a GeoPoint, representing the coordinates of the RoI, generates the requisite mission item, and adds it to the mission queue. This feature was added to facilitate automatic drone navigation around a specific geographical point of interest.

Asana: https://app.asana.com/0/1203851531040615/1205237266324949/f