feat(sdr): add mission functionality to SDR Device
A new functional feature was added to the SDR Device which enables it to conduct missions. The modification was made in several files like SdrServerDevice.cs, ISdrServerDevice.cs, asv_sdr.xml, asv_sdr.cs, SdrClientDevice.cs, ISdrClientDevice.cs.
The main change is the introduction of the 'Missions' object to the SDR device. It helps us to perform different missions using SDR payloads.
The update in asv_sdr.xml file has portrayed how this new feature can be employed for performing tasks in different circumstances. Overall, this enhancement will be helpful to utilize SDR payloads in a much effective manner for mission purposes.