feat (tests): add tests for MissionMicroservice functionalities
In this commit, multiple Unit Tests were added to the MissionMicroserviceTest.cs class. The added tests check the functionality of mission-related features such as setting the current mission, checking if a mission is reached, requesting the mission count, requesting specific mission items, and uploading missions. These tests are crucial to ensuring that the mission microservice functionality is working as expected.

Asana: https://app.asana.com/0/1203851531040615/1205600785681216/f