feat: add SetVehicleSpeed method to IMissionClientEx
A new method named SetVehicleSpeed was added to the IMissionClientEx interface. This method allows to change the speed of the vehicle while it's on a mission. This functionality was necessary to give users more control over the pace of the vehicle during mission execution.

Asana: https://app.asana.com/0/1203851531040615/1205990900963019/f